+++ Assets/Nova/Sources/Scripts/UI/Views/GameViewController.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: de783f267d0570645b528696aa32c5a5
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
